test(adr-derive): move the deriveAdrs tests to the module they now test
CI caught what the module split left behind: `kit check --category tests
--enforce-tests` reported `src/commands/adr-derive.ts` as 1 new untested file. The
tests for `deriveAdrs` were still sitting in `commands/adr.test.ts`, where the code
lived before it moved one module deeper to keep its flags out of review/baseline/
standards. The code migrated; the tests did not.

A move, not an addition — same three tests, same assertions, now in the file whose
name the coverage check derives from the source. Locally the check goes from "1 new
untested file(s) (83 total)" to "82 pre-existing (baseline-frozen)", exit 0.
